+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=parabank
     --cov-report term-missing
 
+[coverage:run]
+source = parabank
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = parabank/locale
 domain = clld
