+++ setup.py
@@ -20,15 +20,20 @@ setup(
     include_package_data=True,
     zip_safe=False,
     install_requires=[
+        'clld~=4.0',
         'clldmpg~=3.1',
         'pyglottolog',
         'clld-glottologfamily-plugin',
+        'sqlalchemy',
+        'waitress',
     ],
     extras_require={
-        'dev': ['flake8', 'waitress'],
-        'test': [
+        'dev': [
+            'flake8',
             'psycopg2',
             'tox',
+        ],
+        'test': [
             'mock',
             'pytest>=3.1',
             'pytest-clld',
