--- sqlite.ini
@@ -1,48 +1,0 @@
-[app:main]
-use = egg:parabank
-
-pyramid.reload_templates = true
-pyramid.debug_authorization = false
-pyramid.debug_notfound = true
-pyramid.debug_routematch = false
-pyramid.default_locale_name = en
-pyramid.includes =
-    pyramid_tm
-sqlalchemy.url = sqlite:///%(here)s/parabank.sqlite
-
-[server:main]
-use = egg:waitress#main
-host = 127.0.0.1
-port = 6543
-
-[loggers]
-keys = root, parabank, sqlalchemy
-
-[handlers]
-keys = console
-
-[formatters]
-keys = generic
-
-[logger_root]
-level = INFO
-handlers = console
-
-[logger_parabank]
-level = DEBUG
-handlers =
-qualname = parabank
-
-[logger_sqlalchemy]
-level = WARN
-handlers =
-qualname = sqlalchemy.engine
-
-[handler_console]
-class = StreamHandler
-args = (sys.stderr,)
-level = NOTSET
-formatter = generic
-
-[formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
