+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest_plugins = ['clld']
