+++ tox.ini
@@ -1,23 +1,11 @@
 [tox]
 envlist =
-    py34,py27
+    py{27,34,35,36}
+skip_missing_interpreters =
+    true
 
 [testenv]
+extras =
+    test
 commands =
-    python -m pytest {posargs}
-deps =
-    pytest
-    mock
-    psycopg2
-    coverage
-    pytest-clld
-    pytest-cov
-    pytest-mock
-    webtest
-
-[testenv:py34]
-basepython = python3.4
-
-[testenv:py27]
-basepython = python2.7
-
+    python -m pytest -m "not selenium" {posargs}
